+++ src/main/java/com/example/demo/services/QuestionnaireService.java
@@ -8,7 +8,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -69,14 +68,36 @@ public class QuestionnaireService {
         }
     }
 
+//      TODO remake delete ResponseEntity<String> -> ResponseEntity<Questionnaire>
 
-    public void deleteQuestionnaire(Long id) {
-        questionnaireRepository.deleteById(id);
+    public ResponseEntity<String> deleteQuestionnaire(Long id) {
+        try {
+            if (questionnaireRepository.findById(id).isEmpty()){
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Questionnaire doesn't exist");
+            }
+            Questionnaire question = questionnaireRepository.findById(id).get();
+            questionnaireRepository.deleteById(id);
+            return ResponseEntity.status(HttpStatus.OK).body(question.toString());
+
+        }catch (Exception e){
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Failed to delete Questionnaire");
+        }
     }
 
-    public Questionnaire getQuestionnaire(Long id) {
-        return questionnaireRepository.findById(id).get();
+
+//      TODO remake get ResponseEntity<String> -> ResponseEntity<Questionnaire>
+    public ResponseEntity<String> getQuestionnaire(Long id) {
+        try {
+            if (questionnaireRepository.findById(id).isPresent()) {
+                return ResponseEntity.status(HttpStatus.OK).body(questionnaireRepository.findById(id).get().toString());
+            }
+        }catch (Exception e){
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Failed to get Questionnaire");
+        }
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Questionnaire isn't found");
+
     }
+
 
     public List<Questionnaire> getActiveQuestionnaires(){
         return questionnaireRepository.findActiveQuestionnaires(LocalDate.now());
